Read upload parameters from MediaWiki:Simplebatchupload-parameters
Parameters are:
* Name of template to be stored as text on initial upload
* Upload comment
* Title line of the Special:BatchUpload page

Each line of MediaWiki:Simplebatchupload-parameters is considered as one set of
parameters. Parameters should be separated by pipes (|). The line to be used is
selected by appending the name of the template to be stored as text on initial
upload as the subpage to the URL of the Special:BatchUpload page.

Example
-------

Consider the parameter line
`Pics | These pics were uploaded using [[mw:Extension:SimpleBatchUpload{{!}}SimpleBatchUpload]] | Upload some pics!`

* This can be selected by going to `Special:BatchUpload/Pics`.
* The title of this page will be `Upload some pics!`.
* The comment for the upload will be `These pics were uploaded using [[mw:Extension:SimpleBatchUpload{{!}}SimpleBatchUpload]]`.
* If a file with that name is uploaded for the first time it will have `{{Pics}}` as wikitext.