Allow to specify the description of all uploaded files.
Before, one could only use an existing template for this via the URL.
This now adds a textarea where you can compose a description just
like for the single file upload, which will be applied to all uploaded
files.

It's backwards compatible since a template passed via the URL will be
placed in the textarea as default content.