Remove id attribute from element that may be repeated on one page.
Due to the fact this function may be called more than once on a page, it might introduce multiple identical textareas on the page. This will cause the <label for=""> to break (clicking on the <label> will simply cause the focus to be on the very first <textarea id="wfUploadDescription"> this way).
Using the label element as a wrapper element, which contains the textarea, will preserve this behaviour for a single upload box, while expanding this behaviour to work properly when multiple upload boxes occur on a single page (see for example https://oldschool.runescape.wiki/w/RuneScape:Batch_Upload).
This will also prevent invalid HTML from being generated by this extension (since re-using the same ID is invalid HTML).

Also indented code to more easily see the HTML structure at a glance.