Fix JS query selector to match the previous id-attribute change
The previous change, while necessary to make the HTML valid, did break
batchuploading with a single instance, due to the id-attribute being used
as a selector to find the textarea. This change will simply look for the
name attribute within the same upload box, in order to match the correct
textarea.

Also updated the release notes with a short description of both this
change and the previous one.