Fix issue breaking batchuploads
Due to one small oversight in my pushed changes last time, the textarea
was not a child of the $(container). This was due to me failing to copy
over all of my tested changes to my editor.

This means that with this small change, everything should be functioning
perfectly. I've tested this more rigorously now and I'm over 99% sure this
works as it should.